pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.6.0",
    "av>=14.3.0",
    "datasets>=3.5.0",
    "deepmultilingualpunctuation>=1.0.1",
    "gradio>=5.25.2",
    "imagehash>=4.3.2",
    "langchain>=0.3.23",
    "langchain-experimental>=0.3.4",
    "llama-index>=0.12.31",
    "loguru>=0.7.3",
    "ollama>=0.4.8",
    "pillow>=11.2.1",
    "pydantic-ai>=0.1.0",
    "pymongo>=4.12.0",
    "qdrant-client>=1.13.3",
    "sentence-transformers>=4.1.0",
    "spacy>=3.8.5",
    "torchvision>=0.21.0",
    "transformers>=4.51.3",
    "youtube-transcript-api>=1.0.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pip>=25.0.1",
    "ruff>=0.11.5",
]
